requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
certifi==2026.1.4
click==8.1.8
dnspython==2.6.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.4
fastapi-cli==0.0.18
fastapi-cloud-cli==0.7.0
fastar==0.8.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.11
jinja2==3.1.6
libsql-experimental==0.0.55
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==26.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.3.2
rich-toolkit==0.19.2
rignore==0.7.6
sentry-sdk==2.52.0
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.46
sqlalchemy-libsql==0.2.0
starlette==0.44.0
tomli==2.4.0
typer==0.20.1
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
